scripts/vision/requirements.txt
tensorflow>=2.11.0
numpy
pillow
opencv-python